Move RepoData to utils module
